# openapi.yaml
# https://editor.swagger.io/
openapi: 3.0.0
info:
  title: DV Video Downloader API
  description: 一个基于 Node.js 的自动化视频下载和处理工具
  version: 1.0.0

servers:
  - url: http://localhost:3456
    description: 本地开发服务器

paths:
  /:
    get:
      summary: 获取欢迎信息
      description: 返回一个简单的欢迎消息
      responses:
        "200":
          description: 成功返回欢迎信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Hello World"

  /download:
    post:
      summary: 提交下载任务
      description: 提交一个视频下载任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: 视频链接
                  example: "https://www.example.com/video"
                quality:
                  type: string
                  description: 视频质量
                  default: "video_best"
                  enum: [video_best, audio_best, video_worst, audio_worst]
                  example: "video_best"
              required:
                - url
      responses:
        "200":
          description: 任务提交成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  taskId:
                    type: string
                    description: 任务ID
                    example: "550e8400-e29b-41d4-a716-446655440000"

  /task/{id}:
    get:
      summary: 获取任务详情
      description: 根据任务ID获取任务的详细信息
      parameters:
        - name: id
          in: path
          required: true
          description: 任务ID
          schema:
            type: string
      responses:
        "200":
          description: 任务信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: 任务不存在

  /c:
    get:
      summary: 测试连接并获取运行中任务
      description: 返回当前正在执行的任务列表
      responses:
        "200":
          description: 连接测试成功，返回运行中的任务列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Connection test successful"
                  runningTasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                  total:
                    type: integer
                    example: 2

  /downloads/{filename}:
    get:
      summary: 获取下载的文件
      description: 获取downloads目录下的文件
      parameters:
        - name: filename
          in: path
          required: true
          description: 文件名
          schema:
            type: string
      responses:
        "200":
          description: 文件内容
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        "404":
          description: 文件不存在

  /stats:
    get:
      summary: 获取任务统计信息
      description: 获取下载任务的统计信息，包括总任务数、正在下载数和失败数
      responses:
        "200":
          description: 任务统计信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    description: 状态码，0表示成功
                    example: 0
                  data:
                    type: object
                    description: 统计数据
                    properties:
                      total:
                        type: integer
                        description: 任务总量
                        example: 100
                      downloading:
                        type: integer
                        description: 正在下载数量（包括pending和running状态）
                        example: 5
                      failed:
                        type: integer
                        description: 下载失败数量
                        example: 3
                  message:
                    type: string
                    description: 响应消息
                    example: "success"


  /tasks/query:
    post:
      summary: 多条件查询任务
      description: 根据多种条件筛选任务，支持所有字段的精确、范围、数组（in）查询，支持分页
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filters:
                  $ref: '#/components/schemas/TaskQueryFilters'
                page:
                  type: integer
                  description: 页码，从1开始
                  default: 1
                limit:
                  type: integer
                  description: 每页数量
                  default: 20
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  pagination:
                    type: object
                    description: 分页信息
                    properties:
                      page:
                        type: integer
                        description: 当前页码
                        example: 1
                      pageSize:
                        type: integer
                        description: 每页大小
                        example: 10
                      total:
                        type: integer
                        description: 总记录数
                        example: 12
                      totalPages:
                        type: integer
                        description: 总页数
                        example: 2
                  message:
                    type: string
                    example: success
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          description: 任务ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        url:
          type: string
          description: 视频URL
          example: "https://www.example.com/video"
        quality:
          type: string
          description: 视频质量
          example: "video_best"
        status:
          type: string
          description: 任务状态
          enum: [pending, running, success, failed]
          example: "running"
        location:
          type: string
          description: 任务执行位置
          example: "local"
        createdAt:
          type: integer
          description: 创建时间戳
          example: 1678886400000
        startedAt:
          type: integer
          description: 开始时间戳
          example: 1678886460000
        finishedAt:
          type: integer
          description: 完成时间戳
          example: 1678886520000
        strategy:
          type: string
          description: 使用的策略
          example: "ytDlpStrategy"
        output:
          type: string
          description: 输出路径
          example: "/path/to/output.mp4"
        outputName:
          type: string
          description: 输出文件名
          example: "output.mp4"
        error:
          type: string
          description: 错误信息（如果有）
          example: "Some error message"
    TaskQueryFilters:
      type: object
      properties:
        id:
          type: string
          description: 任务ID，精确匹配
        url:
          type: string
          description: 视频URL，精确匹配
        quality:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: 视频质量，支持单值或数组（in 查询）
        status:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: 任务状态，支持单值或数组
        location:
          type: string
          description: 任务执行位置
        error:
          type: string
          description: 错误信息
        createdAt:
          oneOf:
            - type: integer
            - type: object
              properties:
                min:
                  type: integer
                  description: 最小创建时间戳
                max:
                  type: integer
                  description: 最大创建时间戳
          description: 创建时间戳，支持范围
        startedAt:
          oneOf:
            - type: integer
            - type: object
              properties:
                min:
                  type: integer
                max:
                  type: integer
          description: 开始时间戳，支持范围
        finishedAt:
          oneOf:
            - type: integer
            - type: object
              properties:
                min:
                  type: integer
                max:
                  type: integer
          description: 完成时间戳，支持范围
        strategy:
          type: string
          description: 策略
        output:
          type: string
          description: 输出路径
        outputName:
          type: string
          description: 输出文件名